Checkpoint
Trying to overcome this hydration problem. I have a suspicion that
I need to do some fairly radical compilation hijinx to get ready-to-go
hydration JS into the browser. This commit includes a fancy route that
slurps up the compiled JS for hydrating a given component. But it doesn't
work (I believe) because the JS isn't compiled for each component. My
intuition is telling me that I need another piece of compilation here
that looks for "hydrate.ts" files and transpiles them into standalone
JS files. That way when the client loads from SSR react, it can also
pull in a single script that hydrates that page. Whew. SSR React is
really tricky. Unclear if this is worth it at all.