Checkpoint: Google OAuth
Working on the Google OAuth handshake. Need a CSRF token. Was trying to
do something fancy, but I think it was kinda dumb. Probably best to just
generate a token and stick it in the DB with any metadata I want to use
for security. That way I'm never exposed any encryption key output at all.