NotificationList: Make it pointerEvents none when not rendered on screen.
Summary:

On page load, MUI renders the NotificationList on the page, but `visibility: hidden`. It rendered it on top of the navigation on the left side of the screen.
I could not use the navigation, and upon inspecting the DOM, found out that yeah.. I was hovering on the notification list.

Fixing by ignoring all pointer events when notification list is closed.

I really wonder why this isn't MUI default?!?