Remove minHeight 100vh from Shell
Summary:
This was unnecessarily adding scrollbars.
To remove this and not introduce regression, I had to make some changes to Nav and TeamChat too.
WIthout those changes, both component would not fill up the entire screen.
Introduced a useful `fullscreen` css rule and set both `max` and `min` height to it. This is because the cord thread is position fixed, so we need to make sure when it grows tall, it says within the viewport.

Test plan:
Loaded a bunch of screens, checked Nav and TeamChat cover entire screen like before.
Check that tall threads don't push the composer offscreen.