Make token swapping "work"
This is some absolutely heinous nonsense. The root of the problem is that I
can't know when <CordProvider /> has initialized / re-initialized without some
even wilder hacks. I need to make sure I don't created thread IDs is the wrong
org, because if I do, everything is ruined. At the moment, my best option is to
just give it a little time to breathe and hope for the best.